delete unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1711876 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/java/org/apache/poi/xssf/streaming/SXSSFRow.java
@@ -372,7 +372,7 @@ public class SXSSFRow implements Row
      */
     public float getHeightInPoints()
     {
-        return (float)(_height==-1?getSheet().getDefaultRowHeightInPoints():(float)_height/20.0);
+        return (float)(_height==-1?getSheet().getDefaultRowHeightInPoints():_height/20.0);
     }
     
     /**
